requirements.txt
fastapi==0.112.1
numpy==2.0.1
opencv-python==4.10.0.84
pillow==10.4.0
pydantic==2.8.2
pytesseract==0.3.10
requests==2.32.3
scipy==1.13.1
selenium==4.23.1
uvicorn